Modifica finale es_17

+++ 11-iterazione/es_17/es_17.c
@@ -6,15 +6,17 @@
  * quattro chicchi di riso per la terza casella... e così via per tutte le 64 caselle della scacchiera>>.
  * Progettare un algoritmo che, a partire dal numero n di caselle che si intendono riempire,
  * calcoli il numero complessivo di chicchi di riso che spettano come ricompensa.
+ *
  * @param argc è il numero di elementi inseriti dalla linea di comando
  * @param argv è un array di stringhe con gli argomenti passati da riga di comando
+ *
  * @return ritorna 0
  */
 int main(int argc, char *argv[])
 {
     int n;
-    unsigned long long chicchiDiRiso = 1;
-    long long chicchiCasella = 1;
+    double chicchiDiRiso = 1;
+    double chicchiCasella = 1;
 
     do
     {
@@ -27,6 +29,6 @@ int main(int argc, char *argv[])
         chicchiDiRiso += chicchiCasella;
         chicchiCasella = chicchiCasella * 2;
     }
-    printf("\nIl numero di chicchi di riso complessivo che spettano come ricompensa è : %llu\n", chicchiDiRiso);
+    printf("\nIl numero di chicchi di riso complessivo che spettano come ricompensa è : %.0f\n", chicchiDiRiso);
     return 0;
 }
